add possibility to use Realtime DB for RawData, but tomorrow I've to test all

+++ app/src/main/java/com/example/chillinapp/data/stress/StressData.kt
@@ -2,8 +2,8 @@ package com.example.chillinapp.data.stress
 
 data class StressRawData(
     val timestamp: Long = 0,
-    val heartrateSensor: Float = 0.0F,
-    val skinTemperatureSensor: Float = 0.0F
+    val heartrateSensor: Double = 0.0,
+    val skinTemperatureSensor: Double = 0.0
 )
 
 data class StressDerivedData (
